Move testTemporality to assertions.js.
